+++ packages/service-core-tests/src/tests/register-data-storage-tests.ts
@@ -34,18 +34,19 @@ const normalizeOplogData = (data: OplogEntry['data']) => {
  */
 export function registerDataStorageTests(generateStorageFactory: storage.TestStorageFactory) {
   test('save and load parameters', async () => {
-    const sync_rules = test_utils.testRules(`
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT group_id FROM test WHERE id1 = token_parameters.user_id OR id2 = token_parameters.user_id
-    data: [] 
-    `);
+    data: []
+    `
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: TEST_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -69,11 +70,12 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('t1')
       });
+
+      await batch.commit('1/1');
     });
 
-    const parameters = await bucketStorage.getParameterSets(result!.flushed_op, [
-      ParameterLookup.normalized('mybucket', '1', ['user1'])
-    ]);
+    const checkpoint = await bucketStorage.getCheckpoint();
+    const parameters = await checkpoint.getParameterSets([ParameterLookup.normalized('mybucket', '1', ['user1'])]);
     expect(parameters).toEqual([
       {
         group_id: 'group1a'
@@ -82,20 +84,19 @@ bucket_definitions:
   });
 
   test('it should use the latest version', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT group_id FROM test WHERE id = token_parameters.user_id
-    data: [] 
+    data: []
     `
-    );
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
-
-    const result1 = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: TEST_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -105,8 +106,10 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('user1')
       });
+      await batch.commit('1/1');
     });
-    const result2 = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    const checkpoint1 = await bucketStorage.getCheckpoint();
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: TEST_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -116,11 +119,11 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('user1')
       });
+      await batch.commit('1/2');
     });
+    const checkpoint2 = await bucketStorage.getCheckpoint();
 
-    const parameters = await bucketStorage.getParameterSets(result2!.flushed_op, [
-      ParameterLookup.normalized('mybucket', '1', ['user1'])
-    ]);
+    const parameters = await checkpoint2.getParameterSets([ParameterLookup.normalized('mybucket', '1', ['user1'])]);
     expect(parameters).toEqual([
       {
         group_id: 'group2'
@@ -128,9 +131,7 @@ bucket_definitions:
     ]);
 
     // Use the checkpoint to get older data if relevant
-    const parameters2 = await bucketStorage.getParameterSets(result1!.flushed_op, [
-      ParameterLookup.normalized('mybucket', '1', ['user1'])
-    ]);
+    const parameters2 = await checkpoint1.getParameterSets([ParameterLookup.normalized('mybucket', '1', ['user1'])]);
     expect(parameters2).toEqual([
       {
         group_id: 'group1'
@@ -139,20 +140,19 @@ bucket_definitions:
   });
 
   test('it should use the latest version after updates', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT id AS todo_id
         FROM todos
         WHERE list_id IN token_parameters.list_id
-    data: [] 
+    data: []
     `
-    );
-
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     const table = test_utils.makeTestTable('todos', ['id', 'list_id']);
 
@@ -176,9 +176,11 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('todo2')
       });
+
+      await batch.commit('1/1');
     });
 
-    const result2 = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       // Update the second todo item to now belong to list 1
       await batch.save({
         sourceTable: table,
@@ -189,12 +191,15 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('todo2')
       });
+
+      await batch.commit('1/1');
     });
 
     // We specifically request the todo_ids for both lists.
     // There removal operation for the association of `list2`::`todo2` should not interfere with the new
     // association of `list1`::`todo2`
-    const parameters = await bucketStorage.getParameterSets(result2!.flushed_op, [
+    const checkpoint = await bucketStorage.getCheckpoint();
+    const parameters = await checkpoint.getParameterSets([
       ParameterLookup.normalized('mybucket', '1', ['list1']),
       ParameterLookup.normalized('mybucket', '1', ['list2'])
     ]);
@@ -210,20 +215,19 @@ bucket_definitions:
   });
 
   test('save and load parameters with different number types', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT group_id FROM test WHERE n1 = token_parameters.n1 and f2 = token_parameters.f2 and f3 = token_parameters.f3
     data: []
     `
-    );
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: TEST_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -236,21 +240,23 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('t1')
       });
+
+      await batch.commit('1/1');
     });
 
     const TEST_PARAMS = { group_id: 'group1' };
 
-    const checkpoint = result!.flushed_op;
+    const checkpoint = await bucketStorage.getCheckpoint();
 
-    const parameters1 = await bucketStorage.getParameterSets(checkpoint, [
+    const parameters1 = await checkpoint.getParameterSets([
       ParameterLookup.normalized('mybucket', '1', [314n, 314, 3.14])
     ]);
     expect(parameters1).toEqual([TEST_PARAMS]);
-    const parameters2 = await bucketStorage.getParameterSets(checkpoint, [
+    const parameters2 = await checkpoint.getParameterSets([
       ParameterLookup.normalized('mybucket', '1', [314, 314n, 3.14])
     ]);
     expect(parameters2).toEqual([TEST_PARAMS]);
-    const parameters3 = await bucketStorage.getParameterSets(checkpoint, [
+    const parameters3 = await checkpoint.getParameterSets([
       ParameterLookup.normalized('mybucket', '1', [314n, 314, 3])
     ]);
     expect(parameters3).toEqual([]);
@@ -261,20 +267,19 @@ bucket_definitions:
     // This specific case tested here cannot happen with postgres in practice, but we still
     // test this to ensure correct deserialization.
 
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT group_id FROM test WHERE n1 = token_parameters.n1
     data: []
     `
-    );
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: TEST_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -298,31 +303,33 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('t1')
       });
+
+      await batch.commit('1/1');
     });
 
     const TEST_PARAMS = { group_id: 'group1' };
 
-    const checkpoint = result!.flushed_op;
+    const checkpoint = await bucketStorage.getCheckpoint();
 
-    const parameters1 = await bucketStorage.getParameterSets(checkpoint, [
+    const parameters1 = await checkpoint.getParameterSets([
       ParameterLookup.normalized('mybucket', '1', [1152921504606846976n])
     ]);
     expect(parameters1).toEqual([TEST_PARAMS]);
   });
 
   test('removing row', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
 
       await batch.save({
@@ -339,9 +346,10 @@ bucket_definitions:
         tag: storage.SaveOperationTag.DELETE,
         beforeReplicaId: test_utils.rid('test1')
       });
+      await batch.commit('1/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
 
     const batch = await test_utils.fromAsync(bucketStorage.getBucketDataBatch(checkpoint, new Map([['global[]', 0n]])));
     const data = batch[0].chunkData.data.map((d) => {
@@ -373,8 +381,9 @@ bucket_definitions:
   test('save and load parameters with workspaceId', async () => {
     const WORKSPACE_TABLE = test_utils.makeTestTable('workspace', ['id']);
 
-    const sync_rules_content = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
     by_workspace:
       parameters:
@@ -382,13 +391,11 @@ bucket_definitions:
           workspace."userId" = token_parameters.user_id
       data: []
     `
-    );
-    const sync_rules = sync_rules_content.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    });
+    const sync_rules = syncRules.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules_content);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: WORKSPACE_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -398,9 +405,9 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('workspace1')
       });
+      await batch.commit('1/1');
     });
-
-    const checkpoint = result!.flushed_op;
+    const checkpoint = await bucketStorage.getCheckpoint();
 
     const parameters = new RequestParameters({ sub: 'u1' }, {});
 
@@ -409,12 +416,12 @@ bucket_definitions:
     const lookups = q1.getLookups(parameters);
     expect(lookups).toEqual([ParameterLookup.normalized('by_workspace', '1', ['u1'])]);
 
-    const parameter_sets = await bucketStorage.getParameterSets(checkpoint, lookups);
+    const parameter_sets = await checkpoint.getParameterSets(lookups);
     expect(parameter_sets).toEqual([{ workspace_id: 'workspace1' }]);
 
     const buckets = await sync_rules.getBucketParameterQuerier(parameters).queryDynamicBucketDescriptions({
       getParameterSets(lookups) {
-        return bucketStorage.getParameterSets(checkpoint, lookups);
+        return checkpoint.getParameterSets(lookups);
       }
     });
     expect(buckets).toEqual([{ bucket: 'by_workspace["workspace1"]', priority: 3 }]);
@@ -423,8 +430,9 @@ bucket_definitions:
   test('save and load parameters with dynamic global buckets', async () => {
     const WORKSPACE_TABLE = test_utils.makeTestTable('workspace');
 
-    const sync_rules_content = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
     by_public_workspace:
       parameters:
@@ -432,13 +440,11 @@ bucket_definitions:
           workspace.visibility = 'public'
       data: []
     `
-    );
-    const sync_rules = sync_rules_content.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    });
+    const sync_rules = syncRules.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules_content);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: WORKSPACE_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -468,9 +474,11 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('workspace3')
       });
+
+      await batch.commit('1/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const checkpoint = await bucketStorage.getCheckpoint();
 
     const parameters = new RequestParameters({ sub: 'unknown' }, {});
 
@@ -479,13 +487,13 @@ bucket_definitions:
     const lookups = q1.getLookups(parameters);
     expect(lookups).toEqual([ParameterLookup.normalized('by_public_workspace', '1', [])]);
 
-    const parameter_sets = await bucketStorage.getParameterSets(checkpoint, lookups);
+    const parameter_sets = await checkpoint.getParameterSets(lookups);
     parameter_sets.sort((a, b) => JSON.stringify(a).localeCompare(JSON.stringify(b)));
     expect(parameter_sets).toEqual([{ workspace_id: 'workspace1' }, { workspace_id: 'workspace3' }]);
 
     const buckets = await sync_rules.getBucketParameterQuerier(parameters).queryDynamicBucketDescriptions({
       getParameterSets(lookups) {
-        return bucketStorage.getParameterSets(checkpoint, lookups);
+        return checkpoint.getParameterSets(lookups);
       }
     });
     buckets.sort((a, b) => a.bucket.localeCompare(b.bucket));
@@ -498,8 +506,9 @@ bucket_definitions:
   test('multiple parameter queries', async () => {
     const WORKSPACE_TABLE = test_utils.makeTestTable('workspace');
 
-    const sync_rules_content = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
     by_workspace:
       parameters:
@@ -509,13 +518,11 @@ bucket_definitions:
             workspace.user_id = token_parameters.user_id
       data: []
     `
-    );
-    const sync_rules = sync_rules_content.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    });
+    const sync_rules = syncRules.parsed(test_utils.PARSE_OPTIONS).sync_rules;
+    const bucketStorage = factory.getInstance(syncRules);
 
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules_content);
-
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable: WORKSPACE_TABLE,
         tag: storage.SaveOperationTag.INSERT,
@@ -557,9 +564,11 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('workspace4')
       });
+
+      await batch.commit('1/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const checkpoint = await bucketStorage.getCheckpoint();
 
     const parameters = new RequestParameters({ sub: 'u1' }, {});
 
@@ -568,7 +577,7 @@ bucket_definitions:
     const lookups1 = q1.getLookups(parameters);
     expect(lookups1).toEqual([ParameterLookup.normalized('by_workspace', '1', [])]);
 
-    const parameter_sets1 = await bucketStorage.getParameterSets(checkpoint, lookups1);
+    const parameter_sets1 = await checkpoint.getParameterSets(lookups1);
     parameter_sets1.sort((a, b) => JSON.stringify(a).localeCompare(JSON.stringify(b)));
     expect(parameter_sets1).toEqual([{ workspace_id: 'workspace1' }]);
 
@@ -576,7 +585,7 @@ bucket_definitions:
     const lookups2 = q2.getLookups(parameters);
     expect(lookups2).toEqual([ParameterLookup.normalized('by_workspace', '2', ['u1'])]);
 
-    const parameter_sets2 = await bucketStorage.getParameterSets(checkpoint, lookups2);
+    const parameter_sets2 = await checkpoint.getParameterSets(lookups2);
     parameter_sets2.sort((a, b) => JSON.stringify(a).localeCompare(JSON.stringify(b)));
     expect(parameter_sets2).toEqual([{ workspace_id: 'workspace3' }]);
 
@@ -584,7 +593,7 @@ bucket_definitions:
     const buckets = (
       await sync_rules.getBucketParameterQuerier(parameters).queryDynamicBucketDescriptions({
         getParameterSets(lookups) {
-          return bucketStorage.getParameterSets(checkpoint, lookups);
+          return checkpoint.getParameterSets(lookups);
         }
       })
     ).map((e) => e.bucket);
@@ -593,20 +602,19 @@ bucket_definitions:
   });
 
   test('changing client ids', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT client_id as id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     const sourceTable = TEST_TABLE;
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
         sourceTable,
         tag: storage.SaveOperationTag.INSERT,
@@ -638,8 +646,10 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('test2')
       });
+
+      await batch.commit('1/1');
     });
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
     const batch = await test_utils.fromAsync(bucketStorage.getBucketDataBatch(checkpoint, new Map([['global[]', 0n]])));
     const data = batch[0].chunkData.data.map((d) => {
       return {
@@ -657,16 +667,16 @@ bucket_definitions:
   });
 
   test('re-apply delete', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
@@ -690,9 +700,11 @@ bucket_definitions:
         tag: storage.SaveOperationTag.DELETE,
         beforeReplicaId: test_utils.rid('test1')
       });
+
+      await batch.commit('1/1');
     });
 
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
 
       await batch.save({
@@ -702,7 +714,7 @@ bucket_definitions:
       });
     });
 
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
 
     const batch = await test_utils.fromAsync(bucketStorage.getBucketDataBatch(checkpoint, new Map([['global[]', 0n]])));
     const data = batch[0].chunkData.data.map((d) => {
@@ -732,16 +744,16 @@ bucket_definitions:
   });
 
   test('re-apply update + delete', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
@@ -785,9 +797,11 @@ bucket_definitions:
         tag: storage.SaveOperationTag.DELETE,
         beforeReplicaId: test_utils.rid('test1')
       });
+
+      await batch.commit('1/1');
     });
 
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
 
       await batch.save({
@@ -815,9 +829,11 @@ bucket_definitions:
         tag: storage.SaveOperationTag.DELETE,
         beforeReplicaId: test_utils.rid('test1')
       });
+
+      await batch.commit('2/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
 
     const batch = await test_utils.fromAsync(bucketStorage.getBucketDataBatch(checkpoint, new Map([['global[]', 0n]])));
 
@@ -850,18 +866,17 @@ bucket_definitions:
   });
 
   test('truncate parameters', async () => {
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   mybucket:
     parameters:
       - SELECT group_id FROM test WHERE id1 = token_parameters.user_id OR id2 = token_parameters.user_id
     data: []
     `
-    );
-
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       await batch.save({
@@ -879,11 +894,9 @@ bucket_definitions:
       await batch.truncate([TEST_TABLE]);
     });
 
-    const { checkpoint } = await bucketStorage.getCheckpoint();
+    const checkpoint = await bucketStorage.getCheckpoint();
 
-    const parameters = await bucketStorage.getParameterSets(checkpoint, [
-      ParameterLookup.normalized('mybucket', '1', ['user1'])
-    ]);
+    const parameters = await checkpoint.getParameterSets([ParameterLookup.normalized('mybucket', '1', ['user1'])]);
     expect(parameters).toEqual([]);
   });
 
@@ -896,16 +909,16 @@ bucket_definitions:
     // 1. Not getting the correct "current_data" state for each operation.
     // 2. Output order not being correct.
 
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "test"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     // Pre-setup
     const result1 = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
@@ -1048,14 +1061,6 @@ bucket_definitions:
   });
 
   test('changed data with replica identity full', async () => {
-    const sync_rules = test_utils.testRules(
-      `
-bucket_definitions:
-  global:
-    data:
-      - SELECT id, description FROM "test"
-`
-    );
     function rid2(id: string, description: string) {
       return getUuidReplicaIdentityBson({ id, description }, [
         { name: 'id', type: 'VARCHAR', typeId: 25 },
@@ -1063,7 +1068,15 @@ bucket_definitions:
       ]);
     }
     await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    const syncRules = await factory.updateSyncRules({
+      content: `
+bucket_definitions:
+  global:
+    data:
+      - SELECT id, description FROM "test"
+`
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     const sourceTable = test_utils.makeTestTable('test', ['id', 'description']);
 
@@ -1155,14 +1168,6 @@ bucket_definitions:
   });
 
   test('unchanged data with replica identity full', async () => {
-    const sync_rules = test_utils.testRules(
-      `
-bucket_definitions:
-  global:
-    data:
-      - SELECT id, description FROM "test"
-`
-    );
     function rid2(id: string, description: string) {
       return getUuidReplicaIdentityBson({ id, description }, [
         { name: 'id', type: 'VARCHAR', typeId: 25 },
@@ -1171,7 +1176,15 @@ bucket_definitions:
     }
 
     await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    const syncRules = await factory.updateSyncRules({
+      content: `
+bucket_definitions:
+  global:
+    data:
+      - SELECT id, description FROM "test"
+`
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     const sourceTable = test_utils.makeTestTable('test', ['id', 'description']);
 
@@ -1260,18 +1273,18 @@ bucket_definitions:
     // but large enough in size to be split over multiple returned batches.
     // The specific batch splits is an implementation detail of the storage driver,
     // and the test will have to updated when other implementations are added.
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
 
       const largeDescription = '0123456789'.repeat(12_000_00);
@@ -1316,9 +1329,11 @@ bucket_definitions:
         },
         afterReplicaId: test_utils.rid('test3')
       });
+
+      await batch.commit('1/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
 
     const options: storage.BucketDataBatchOptions = {
       chunkLimitBytes: 16 * 1024 * 1024
@@ -1367,18 +1382,18 @@ bucket_definitions:
 
   test('long batch', async () => {
     // Test syncing a batch of data that is limited by count.
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
 bucket_definitions:
   global:
     data:
       - SELECT id, description FROM "%"
 `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
-    const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+    await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
       const sourceTable = TEST_TABLE;
 
       for (let i = 1; i <= 6; i++) {
@@ -1392,9 +1407,11 @@ bucket_definitions:
           afterReplicaId: `test${i}`
         });
       }
+
+      await batch.commit('1/1');
     });
 
-    const checkpoint = result!.flushed_op;
+    const { checkpoint } = await bucketStorage.getCheckpoint();
 
     const batch1 = await test_utils.oneFromAsync(
       bucketStorage.getBucketDataBatch(checkpoint, new Map([['global[]', 0n]]), { limit: 4 })
@@ -1441,8 +1458,9 @@ bucket_definitions:
 
   describe('batch has_more', () => {
     const setup = async (options: BucketDataBatchOptions) => {
-      const sync_rules = test_utils.testRules(
-        `
+      await using factory = await generateStorageFactory();
+      const syncRules = await factory.updateSyncRules({
+        content: `
   bucket_definitions:
     global1:
       data:
@@ -1451,11 +1469,10 @@ bucket_definitions:
       data:
         - SELECT id, description FROM test WHERE bucket = 'global2'
   `
-      );
-      await using factory = await generateStorageFactory();
-      const bucketStorage = factory.getInstance(sync_rules);
+      });
+      const bucketStorage = factory.getInstance(syncRules);
 
-      const result = await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
+      await bucketStorage.startBatch(test_utils.BATCH_OPTIONS, async (batch) => {
         const sourceTable = TEST_TABLE;
 
         for (let i = 1; i <= 10; i++) {
@@ -1470,9 +1487,11 @@ bucket_definitions:
             afterReplicaId: `test${i}`
           });
         }
+
+        await batch.commit('1/1');
       });
 
-      const checkpoint = result!.flushed_op;
+      const { checkpoint } = await bucketStorage.getCheckpoint();
       return await test_utils.fromAsync(
         bucketStorage.getBucketDataBatch(
           checkpoint,
@@ -1612,8 +1631,9 @@ bucket_definitions:
   });
 
   test('invalidate cached parsed sync rules', async () => {
-    const sync_rules_content = test_utils.testRules(
-      `
+    await using bucketStorageFactory = await generateStorageFactory();
+    const syncRules = await bucketStorageFactory.updateSyncRules({
+      content: `
 bucket_definitions:
     by_workspace:
       parameters:
@@ -1621,10 +1641,8 @@ bucket_definitions:
           workspace."userId" = token_parameters.user_id
       data: []
     `
-    );
-
-    await using bucketStorageFactory = await generateStorageFactory();
-    const syncBucketStorage = bucketStorageFactory.getInstance(sync_rules_content);
+    });
+    const syncBucketStorage = bucketStorageFactory.getInstance(syncRules);
 
     const parsedSchema1 = syncBucketStorage.getParsedSyncRules({
       defaultSchema: 'public'
@@ -1653,7 +1671,7 @@ bucket_definitions:
       content: `
 bucket_definitions:
   mybucket:
-    data: [] 
+    data: []
     `,
       validate: false
     });
@@ -1913,17 +1931,17 @@ bucket_definitions:
     // Test syncing a batch of data that is small in count,
     // but large enough in size to be split over multiple returned chunks.
     // Similar to the above test, but splits over 1MB chunks.
-    const sync_rules = test_utils.testRules(
-      `
+    await using factory = await generateStorageFactory();
+    const syncRules = await factory.updateSyncRules({
+      content: `
   bucket_definitions:
     global:
       data:
         - SELECT id FROM test
         - SELECT id FROM test_ignore WHERE false
   `
-    );
-    await using factory = await generateStorageFactory();
-    const bucketStorage = factory.getInstance(sync_rules);
+    });
+    const bucketStorage = factory.getInstance(syncRules);
 
     const sourceTable = test_utils.makeTestTable('test', ['id']);
     const sourceTableIgnore = test_utils.makeTestTable('test_ignore', ['id']);