+++ packages/service-core-tests/src/tests/tests-index.ts
@@ -1,5 +1,6 @@
 export * from './register-bucket-validation-tests.js';
 export * from './register-compacting-tests.js';
+export * from './register-parameter-compacting-tests.js';
 export * from './register-data-storage-tests.js';
 export * from './register-migration-tests.js';
 export * from './register-sync-tests.js';